doc_loader/requirements.txt
langchain
langchain-groq
langchain-community
pypdf
beautifulsoup4
lxml
python-dotenv
requests
